backend/requirements.txt
django==5.2
djangorestframework==3.16.0
python-dotenv==1.1.0
django-cors-headers==4.7.0
mysqlclient==2.2.7
pillow==11.2.1 